requirements.txt
requests==2.31.0
python-telegram-bot==13.7
wkhtmltopdf==0.2
